Fix missing #include <cstdint> by avoiding the need for it

+++ test/span.t.cpp
@@ -1060,9 +1060,9 @@ CASE( "make_span(): Allows building from a const container (with_container_t, st
 
 CASE( "[.issue 3]" )
 {
-    static const uint8_t data[] = { 1, 2, 3, 4, 5, 6, 7, 8, 9, };
+    static const int data[] = { 1, 2, 3, 4, 5, 6, 7, 8, 9, };
 
-    span< const uint8_t > spn( data );
+    span< const int > spn( data );
 
     EXPECT( spn[0] == 0 );  // suppress: unused parameter 'lest_env' [-Wunused-parameter]
